+++ rollup.config.js
@@ -1,4 +1,8 @@
+import clear from 'rollup-plugin-clear';
 import cleanup from 'rollup-plugin-cleanup';
+import copy from 'rollup-plugin-copy';
+import resolve from 'rollup-plugin-node-resolve';
+import commonJS from 'rollup-plugin-commonjs';
 
 export default {
     input: 'src/lib.js',
@@ -7,6 +11,27 @@ export default {
         format: 'esm'
     },
     plugins: [
-        cleanup()
+        clear({
+            targets: ['rls'],
+        }),
+        resolve(),
+        commonJS({
+              include: 'node_modules/**'
+        }),
+        cleanup(),
+        copy({
+            targets: [
+                {
+                    src: [
+                        'src/assets',
+                        'src/db',
+                        'src/templates',
+                        'src/log',
+                        'src/api.js'
+                    ], 
+                    dest: 'lib'
+                }
+            ]
+        })
     ]
 };