+++ src/lib.js
@@ -1,5 +1,3 @@
 "use strict";
 
-export { pos } from "./components/pos_func.js";
-export { scale } from "./components/scale_func.js";
-export { xy } from "./components/xy_func.js";
+export { scale } from "./scale_func.js";